+++ stroom-query-testing/src/integration-test/java/stroom/query/testing/hibernate/TestHibernateQueryResourceIT.java
@@ -69,6 +69,7 @@ public class TestHibernateQueryResourceIT extends QueryResourceIT<TestDocRefHibe
                                 .extractValues(false)
                                 .showDetail(false)
                                 .addFields(new Field.Builder()
+                                        .id(TestQueryableHibernateEntity.FLAVOUR)
                                         .name(TestQueryableHibernateEntity.FLAVOUR)
                                         .expression("${" + TestQueryableHibernateEntity.FLAVOUR + "}")
                                         .build())